npgsql v-5.0.7 merge

+++ npgsql-5.0.1.1/src/Npgsql.NodaTime/IntervalHandler.cs
@@ -55,6 +55,8 @@ namespace EnterpriseDB.EDBClient.NodaTime
 
         public override void Write(Period value, EDBWriteBuffer buf, EDBParameter? parameter)
         {
+            // Note that the end result must be long
+            // see #3438
             var microsecondsInDay =
                 (((value.Hours * NodaConstants.MinutesPerHour + value.Minutes) * NodaConstants.SecondsPerMinute + value.Seconds) * NodaConstants.MillisecondsPerSecond + value.Milliseconds) * 1000 +
                 value.Nanoseconds / 1000; // Take the microseconds, discard the nanosecond remainder
@@ -80,8 +82,10 @@ namespace EnterpriseDB.EDBClient.NodaTime
 
         public void Write(Duration value, EDBWriteBuffer buf, EDBParameter? parameter)
         {
-            const int microsecondsPerSecond = 1_000_000;
+            const long microsecondsPerSecond = 1_000_000;
 
+            // Note that the end result must be long
+            // see #3438
             var microsecondsInDay =
                 (((value.Hours * NodaConstants.MinutesPerHour + value.Minutes) * NodaConstants.SecondsPerMinute + value.Seconds) *
                     microsecondsPerSecond + value.SubsecondNanoseconds / 1000); // Take the microseconds, discard the nanosecond remainder
